# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Poufmaker API",
    "version": "1.0.0",
    "description": "API documentation for Poufmaker application",
    "contact": {
      "name": "API Support",
      "email": "support@poufmaker.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string"
          },
          "Email": {
            "type": "string",
            "format": "email"
          },
          "FullName": {
            "type": "string"
          },
          "PhoneNumber": {
            "type": "string"
          },
          "Role": {
            "type": "string",
            "enum": [
              "client",
              "admin",
              "upholsterer"
            ]
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string",
            "format": "uuid"
          },
          "Title": {
            "type": "string"
          },
          "Description": {
            "type": "string"
          },
          "Price": {
            "type": "number",
            "format": "float"
          },
          "ImageUrl": {
            "type": "string",
            "format": "uri"
          },
          "Status": {
            "type": "string",
            "enum": [
              "ai-generated",
              "pending",
              "approved",
              "rejected"
            ],
            "default": "ai-generated"
          },
          "CreatorId": {
            "type": "string",
            "format": "uuid"
          },
          "ManufacturerId": {
            "type": "string",
            "format": "uuid"
          },
          "CreatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "UpdatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "users_products_CreatorIdTousers": {
            "$ref": "#/components/schemas/User"
          },
          "users_products_ManufacturerIdTousers": {
            "$ref": "#/components/schemas/User"
          },
          "bids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Bid"
            }
          }
        }
      },
      "Bid": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string",
            "format": "uuid"
          },
          "ProductId": {
            "type": "string",
            "format": "uuid"
          },
          "UpholstererId": {
            "type": "string",
            "format": "uuid"
          },
          "Amount": {
            "type": "number",
            "format": "float"
          },
          "Status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "rejected"
            ],
            "default": "pending"
          },
          "Notes": {
            "type": "string"
          },
          "CreatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "UpdatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get product details",
        "description": "Get detailed information about a specific product",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Product ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Product details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update product",
        "description": "Update an existing product",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Product ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Title": {
                    "type": "string",
                    "example": "Updated Modern Pouf"
                  },
                  "Description": {
                    "type": "string",
                    "example": "An updated comfortable modern pouf"
                  },
                  "Price": {
                    "type": "number",
                    "format": "float",
                    "example": 129.99
                  },
                  "ImageUrl": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://example.com/updated-pouf.jpg"
                  },
                  "Status": {
                    "type": "string",
                    "enum": [
                      "ai-generated",
                      "pending",
                      "approved",
                      "rejected"
                    ]
                  },
                  "ManufacturerId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - user does not have permission"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete product",
        "description": "Delete an existing product",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Product ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - user does not have permission"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}